add new exercise mockup
adds a new binary file `04` and its corresponding main.rs file. `04` contains a function `max_area` that computes the maximum area of a container given a list of heights. an example and the constraints of the problem are also included. furthermore, a unit test is added to ensure the correctness of the implementation.